Two function created in JS file
checkCardFlag will check the cardWaiting flag wether its value is true
or not and toggle.

cardWaiting Flag is used to let game wait for user to choose one more
card to compare with the previous one.

If user select two different cards then checkCardValues get called and
it will fetch the card values and compare them.

If those values are same then it will reset the basic variables or it
will flip the cards back if they are not matching with each other.