Player won the game or not implemented
checkWin() gets executed whenever user selected one card and it will
find the count of flipped card.

If the result count reaches 16 then it will notify used that he won the
game or else game goes on.

It asks user on win that he wants to play the game again or close the
window?